Add user profile view and access control
Implemented a feature that allows users to view other users' profiles. If the user is viewing their own profile, they have access to edit it. If they are viewing another user's profile, they have the option to send a message or add the user to their friends list. This enhances user interaction and personalization on the platform.